IPA-6789: Create NuGet build test

+++ tools/package.sh
@@ -94,23 +94,25 @@ mkdir $ARTIFACT_PATH
 
 
 # Build Python Wheels for a range of Python Versions
-for PYBUILD in `ls -1 /opt/python`; do
-    PYTHON_BIN=/opt/python/${PYBUILD}/bin
-    if [[ "$PYBUILD" == cp26* ]]; then
-        continue
-    fi
-    if [[ "$PYBUILD" == cp33* ]]; then
-        continue
-    fi
-    touch $SOURCE_PATH/src/ext/python/CMakeLists.txt
-    run "Configure ${PYBUILD}" cmake $SOURCE_PATH -B${BUILD_PATH} -DPYTHON_EXECUTABLE=${PYTHON_BIN}/python ${CMAKE_EXTRA_FLAGS} -DSKIP_PACKAGE_ALL_WHEEL=ON -DPYTHON_WHEEL_PREFIX=${ARTIFACT_PATH}/tmp
+if [ -e /opt/python ] ; then
+    for PYBUILD in `ls -1 /opt/python`; do
+        PYTHON_BIN=/opt/python/${PYBUILD}/bin
+        if [[ "$PYBUILD" == cp26* ]]; then
+            continue
+        fi
+        if [[ "$PYBUILD" == cp33* ]]; then
+            continue
+        fi
+        touch $SOURCE_PATH/src/ext/python/CMakeLists.txt
+        run "Configure ${PYBUILD}" cmake $SOURCE_PATH -B${BUILD_PATH} -DPYTHON_EXECUTABLE=${PYTHON_BIN}/python ${CMAKE_EXTRA_FLAGS} -DSKIP_PACKAGE_ALL_WHEEL=ON -DPYTHON_WHEEL_PREFIX=${ARTIFACT_PATH}/tmp
 
-    run "Test ${PYBUILD}" cmake --build $BUILD_PATH --target check -- -j${THREAD_COUNT}
-    run "Build ${PYBUILD}" cmake --build $BUILD_PATH --target package_wheel -- -j${THREAD_COUNT}
-    auditwheel show ${ARTIFACT_PATH}/tmp/interop*${PYBUILD}*linux_x86_64.whl
-    auditwheel repair ${ARTIFACT_PATH}/tmp/interop*${PYBUILD}*linux_x86_64.whl -w ${ARTIFACT_PATH}
-    rm -fr ${ARTIFACT_PATH}/tmp
-done
+        run "Test ${PYBUILD}" cmake --build $BUILD_PATH --target check -- -j${THREAD_COUNT}
+        run "Build ${PYBUILD}" cmake --build $BUILD_PATH --target package_wheel -- -j${THREAD_COUNT}
+        auditwheel show ${ARTIFACT_PATH}/tmp/interop*${PYBUILD}*linux_x86_64.whl
+        auditwheel repair ${ARTIFACT_PATH}/tmp/interop*${PYBUILD}*linux_x86_64.whl -w ${ARTIFACT_PATH}
+        rm -fr ${ARTIFACT_PATH}/tmp
+    done
+fi
 
 if [ ! -z $PYTHON_VERSION ] ; then
     if [ "$PYTHON_VERSION" == "ALL" ] ; then
@@ -161,7 +163,7 @@ fi
 run "Package" cmake --build $BUILD_PATH --target bundle -- -j${THREAD_COUNT}
 
 if hash dotnet 2> /dev/null; then
-    run "Configure DotNetCore" cmake $SOURCE_PATH -B${BUILD_PATH} ${CMAKE_EXTRA_FLAGS} -DCSBUILD_TOOL=DotNetCore && cmake --$BUILD_PATH --target nupack -- -j${THREAD_COUNT} || true
+    run "Configure DotNetCore" cmake $SOURCE_PATH -B${BUILD_PATH} ${CMAKE_EXTRA_FLAGS} -DCSBUILD_TOOL=DotNetCore && cmake --build $BUILD_PATH --target nupack -- -j${THREAD_COUNT} || true
 fi
 rm -fr ${ARTIFACT_PATH}/tmp
 echo "List Artifacts:"
